+++ src/Guard.ts
@@ -224,9 +224,9 @@ declare module 'fp-ts/lib/HKT' {
 }
 
 /**
- * @since 2.2.0
+ * @since 2.2.3
  */
-export const guard: Schemable1<URI> & WithUnknownContainers1<URI> & WithUnion1<URI> & WithRefinement1<URI> = {
+export const schemableGuard: Schemable1<URI> & WithUnknownContainers1<URI> & WithUnion1<URI> & WithRefinement1<URI> = {
   URI,
   literal,
   string,