+++ test/Guard.ts
@@ -187,7 +187,7 @@ describe('Guard', () => {
       b: Array<A>
     }
 
-    const guard: G.Guard<A> = G.guard.lazy('A', () =>
+    const guard: G.Guard<A> = G.schemableGuard.lazy('A', () =>
       G.type({
         a: G.number,
         b: G.array(guard)