+++ src/index.ts
@@ -1554,12 +1554,7 @@ export class ReadonlyType<C extends Any, A = any, O = A, I = unknown> extends Ty
  * @since 1.5.3
  */
 export interface ReadonlyC<C extends Mixed>
-  extends ReadonlyType<
-    C,
-    { readonly [K in keyof TypeOf<C>]: TypeOf<C>[K] },
-    { readonly [K in keyof OutputOf<C>]: OutputOf<C>[K] },
-    unknown
-  > {}
+  extends ReadonlyType<C, Readonly<TypeOf<C>>, Readonly<OutputOf<C>>, unknown> {}
 
 /**
  * @since 1.0.0